Extract `CodeGenerator` into `generation.py`
